Fold the event handler helpers into the host hook file
Move SUPPORTED_EVENTS, EVENT_TYPES and getEventHandlers into the host hook and
delete the now thin editable-root-event-handlers module. The registration hook
imports getEventHandlers from the host file, keeping the feature in one place.
